next-app/src/components/layout/Navigation.tsx
'use client'

import { useState } from 'react'

interface Tab {
  id: string
  label: string
  icon: JSX.Element
}

const tabs: Tab[] = [
  {
    id: 'llm',
    label: 'LLM Assistant',
    icon: (
      <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5" viewBox="0 0 20 20" fill="currentColor">
        <path fillRule="evenodd" d="M18 10a8 8 0 11-16 0 8 8 0 0116 0zm-8-3a1 1 0 00-.867.5 1 1 0 11-1.731-1A3 3 0 0113 8a3.001 3.001 0 01-2 2.83V11a1 1 0 11-2 0v-1a1 1 0 011-1 1 1 0 100-2zm0 8a1 1 0 100-2 1 1 0 000 2z" clipRule="evenodd" />
      </svg>
    ),
  },
  {
    id: 'lexical',
    label: 'Lexical Values',
    icon: (
      <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5" viewBox="0 0 20 20" fill="currentColor">
        <path d="M7 4a1 1 0 011-1h4a1 1 0 011 1v1h3a1 1 0 011 1v10a1 1 0 01-1 1H4a1 1 0 01-1-1V6a1 1 0 011-1h3V4z" />
      </svg>
    ),
  },
  {
    id: 'corpus',
    label: 'Corpus Manager',
    icon: (
      <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5" viewBox="0 0 20 20" fill="currentColor">
        <path d="M9 4.804A7.968 7.968 0 005.5 4c-1.255 0-2.443.29-3.5.804v10A7.969 7.969 0 015.5 14c1.669 0 3.218.51 4.5 1.385A7.962 7.962 0 0114.5 14c1.255 0 2.443.29 3.5.804v-10A7.968 7.968 0 0014.5 4c-1.255 0-2.443.29-3.5.804V12a1 1 0 11-2 0V4.804z" />
      </svg>
    ),
  },
]

interface NavigationProps {
  activeTab: string
  onTabChange: (tabId: string) => void
}

export function Navigation({ activeTab, onTabChange }: NavigationProps) {
  return (
    <div className="tabs tabs-boxed justify-center bg-base-100 p-2 rounded-lg shadow-md">
      {tabs.map((tab) => (
        <button
          key={tab.id}
          className={`tab tab-lg gap-2 ${activeTab === tab.id ? 'tab-active' : ''}`}
          onClick={() => onTabChange(tab.id)}
        >
          {tab.icon}
          {tab.label}
        </button>
      ))}
    </div>
  )
}
